docs: Add TypeScriptEditor to aid type demos (#2355)
