+++ docs/core/getting-started/README.md
@@ -4,3 +4,7 @@
 
 - [Introduction](../README.md)
 - [Installation](./installation.md)
+- [Define Methods](./endpoint.md)
+- [Render Data](./data-dependency.md)
+- [Mutate Data](./mutations.md)
+- [Comparison](./comparison.md)
