+++ website/src/components/Playground/PlaygroundEditor.tsx
@@ -1,5 +1,5 @@
-import PlaygroundMonacoEditor from './PlaygroundMonacoEditor';
 import PlaygroundLiveEditor from './PlaygroundLiveEditor';
+import PlaygroundMonacoEditor from './PlaygroundMonacoEditor';
 import usingMonaco from './usingMonaco';
 
 const PlaygroundEditor = usingMonaco
